[fuzzing] Add fuzzing.gni
This CL adds the GN rules to create fuzzers.  It will build a libFuzzer
style binary correctly, and bundle it along with needed data files.

Fuzzers are split into fuzz targets, which are the libFuzzer
executables, and fuzz packages, which put fuzz targets into Fuchsia
packages. The fuzz targets will ONLY be built if it is included in a
fuzz package that specifies a sanitizer that corresponds to the fuzz
target's select_variant.  The fuzz package will be built regardless
of select_variant settings, but may be vacuous.

SEC-102 #done

Change-Id: I3a59ac118ac6c39f3c0d1bfa8ae9d9f1056502da